+++ components/src/core/drawer/drawer-body/nav-item/drawer-nav-item.module.ts
@@ -6,18 +6,18 @@ import { MatIconModule } from '@angular/material/icon';
 import { MatDividerModule } from '@angular/material/divider';
 import { MatRippleModule } from '@angular/material/core';
 import { MatExpansionModule } from '@angular/material/expansion';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { MatTooltipModule } from '@angular/material/tooltip';
 
 @NgModule({
     declarations: [DrawerNavItemComponent],
     imports: [
+        MatTooltipModule,
         CommonModule,
         InfoListItemModule,
         MatIconModule,
         MatDividerModule,
         MatRippleModule,
         MatExpansionModule,
-        BrowserAnimationsModule,
     ],
     exports: [DrawerNavItemComponent],
 })